requirements.txt
affine==2.3.0
attrs==20.1.0
certifi==2020.6.20
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
opencv-python==4.4.0.42
openpyxl==3.0.5
pandas==1.1.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
rasterio==1.1.5
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
snuggs==1.4.7
spectral==0.22.1
threadpoolctl==2.1.0
Werkzeug==1.0.1
WTForms==2.3.3

